Fixed player bugs,not showing videos
